+++ src/components/App/AppUI.js
@@ -7,39 +7,50 @@ import { TodosLoading } from '../TodosLoading/TodosLoading';
 import { TodosError } from '../TodosError/TodosError';
 import { EmptyTodos } from '../EmptyTodos/EmptyTodos';
 import React from 'react';
+import { TodoContext } from '../TodoContext/TodoContext';
 
 function APPUI({
-  todoCompleted,
-  totalTodos,
-  searchValue,
-  setSearchValue,
-  searchedTodos,
-  completeTodo,
-  deleteTodo,
-  loading,
-  error
+  // todoCompleted,
+  // totalTodos,
+  // searchValue,
+  // setSearchValue,
+  // searchedTodos,
+  // completeTodo,
+  // deleteTodo,
+  // loading,
+  // error
 }) {
   return (
     <React.Fragment>
-      <TodoCounter completed={todoCompleted} total={totalTodos} />
-      <TodoSearch searchValue={searchValue} setSearchValue={setSearchValue} />
+      
+      <TodoCounter />
+      <TodoSearch  />
 
-      <TodoList>
-        {loading && <TodosLoading /> }
-        {error && <TodosError />}
-        { ( !loading && searchedTodos.length === 0   ) && <EmptyTodos /> }
+      <TodoContext.Consumer>
+        {({
+            searchedTodos,
+            completeTodo,
+            deleteTodo,
+            loading,
+            error
+        }) => (
+          <TodoList>
+            {loading && <TodosLoading /> }
+            {error && <TodosError />}
+            { ( !loading && searchedTodos.length === 0   ) && <EmptyTodos /> }
 
-        {searchedTodos.map((todo) => (
-          <TodoItem
-            key={todo.text}
-            text={todo.text}
-            completed={todo.completed}
-            onCompleted={() => completeTodo(todo.text)}
-            onDeleted={() => deleteTodo(todo.text)}
-          />
-        ))}
-      </TodoList>
-
+            {searchedTodos.map((todo) => (
+              <TodoItem
+                key={todo.text}
+                text={todo.text}
+                completed={todo.completed}
+                onCompleted={() => completeTodo(todo.text)}
+                onDeleted={() => deleteTodo(todo.text)}
+              />
+            ))}
+        </TodoList>
+        )}
+      </TodoContext.Consumer>
       <CreateTodoButton />
     </React.Fragment>
   );
