+++ src/components/TodoCounter/TodoCounter.js
@@ -1,9 +1,16 @@
+import React from 'react';
 import './TodoCounter.css';
+import { TodoContext } from '../TodoContext/TodoContext';
 
 function TodoCounter({ total, completed }){
+    const {
+        todoCompleted,
+        totalTodos,
+    } = React.useContext(TodoContext);
+
     return(
         <h1 className='TodoCounter'>
-            Has completado <span> { completed } </span> de <span> { total } </span> TODOs
+            Has completado <span> { todoCompleted } </span> de <span> { totalTodos } </span> TODOs
         </h1>
     );
 }
