+++ src/components/TodoSearch/TodoSearch.js
@@ -1,7 +1,15 @@
 // import React, { useState } from 'react';
+import React from 'react';
 import './TodoSearch.css';
+import { TodoContext } from '../TodoContext/TodoContext';
 
-function TodoSearch({ searchValue, setSearchValue}){
+function TodoSearch(){
+
+  const {
+    searchValue, 
+    setSearchValue
+} = React.useContext(TodoContext);
+
 
     return(
